Revert "Change waiting strategy for PG imgs from RH registry till FW fixed"
This reverts commit c4ca78c3bf0958fc59c1af849092afac367cd211.
