+++ external-applications/src/test/java/io/quarkus/ts/external/applications/TodoDemoIT.java
@@ -1,13 +1,24 @@
 package io.quarkus.ts.external.applications;
 
+import static org.hamcrest.Matchers.is;
+
+import org.apache.http.HttpStatus;
+import org.junit.jupiter.api.Test;
+
 import io.quarkus.test.bootstrap.PostgresqlService;
 import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.scenarios.annotations.DisabledOnNative;
 import io.quarkus.test.services.Container;
 import io.quarkus.test.services.GitRepositoryQuarkusApplication;
+import io.restassured.http.ContentType;
 
+@DisabledOnNative(reason = "This scenario is using uber-jar, so it's incompatible with Native")
 @QuarkusScenario
-public class TodoDemoIT extends AbstractTodoDemoIT {
+public class TodoDemoIT {
+    private static final String TODO_REPO = "https://github.com/quarkusio/todo-demo-app.git";
+    private static final String VERSIONS = "-Dquarkus.platform.group-id=${QUARKUS_PLATFORM_GROUP-ID} -Dquarkus.platform.version=${QUARKUS_PLATFORM_VERSION} ";
+    private static final String DEFAULT_OPTIONS = " -DskipTests=true " + VERSIONS;
 
     @Container(image = "${postgresql.latest.image}", port = 5432, expectedLog = "listening on IPv4 address")
     static PostgresqlService database = new PostgresqlService()
@@ -27,13 +38,24 @@ public class TodoDemoIT extends AbstractTodoDemoIT {
             .withProperty("quarkus.datasource.password", database.getPassword())
             .withProperty("quarkus.datasource.jdbc.url", database::getJdbcUrl);
 
-    @Override
-    protected RestService getApp() {
-        return app;
+    @Test
+    public void startsSuccessfully() {
+        app.given()
+                .contentType(ContentType.JSON)
+                .body("{\"title\": \"Use Quarkus\", \"order\": 1, \"url\": \"https://quarkus.io\"}")
+                .post("/api")
+                .then()
+                .statusCode(HttpStatus.SC_CREATED);
     }
 
-    @Override
-    protected RestService getReplaced() {
-        return replaced;
+    @Test
+    public void replacedStartsSuccessfully() {
+        replaced.given()
+                .accept(ContentType.JSON)
+                .get("/api")
+                .then()
+                .statusCode(HttpStatus.SC_OK)
+                .body("$.size()", is(1))
+                .body("title[0]", is("Use Quarkus"));
     }
 }
