+++ sql-db/reactive-rest-data-panache/src/test/java/io/quarkus/ts/reactive/rest/data/panache/PostgresqlPanacheResourceIT.java
@@ -9,7 +9,7 @@ import io.quarkus.test.services.QuarkusApplication;
 @QuarkusScenario
 public class PostgresqlPanacheResourceIT extends AbstractPanacheResourceIT {
 
-    private static final int POSTGRESQL_PORT = 5432;
+    static final int POSTGRESQL_PORT = 5432;
 
     @Container(image = "${postgresql.latest.image}", port = POSTGRESQL_PORT, expectedLog = "listening on IPv4 address")
     public static final PostgresqlService database = new PostgresqlService()
@@ -21,5 +21,4 @@ public class PostgresqlPanacheResourceIT extends AbstractPanacheResourceIT {
             .withProperty("quarkus.datasource.username", database.getUser())
             .withProperty("quarkus.datasource.password", database.getPassword())
             .withProperty("quarkus.datasource.jdbc.url", database::getJdbcUrl);
-
 }
